[Feat] 온보딩 첫 번째 화면 위젯으로 분리 (#119)
