+++ lib/common/bouncing.dart
@@ -1,10 +1,12 @@
+// ignore_for_file: library_private_types_in_public_api
+
 import 'package:flutter/material.dart';
 
 class Bouncing extends StatefulWidget {
   final Widget child;
   final Function()? onPress;
 
-  Bouncing({required this.child, Key? key, this.onPress}) : super(key: key);
+  const Bouncing({required this.child, Key? key, this.onPress}) : super(key: key);
 
   @override
   _BouncingState createState() => _BouncingState();
@@ -20,7 +22,7 @@ class _BouncingState extends State<Bouncing>
     super.initState();
     _controller = AnimationController(
       vsync: this,
-      duration: Duration(milliseconds: 100),
+      duration: const Duration(milliseconds: 100),
       lowerBound: 0.0,
       upperBound: 0.05,
     );
