--- lib/common/custom_app_bar.dart
@@ -1,21 +1,0 @@
-import 'package:flutter/material.dart';
-
-class CustomAppBar extends StatelessWidget implements PreferredSizeWidget {
-  const CustomAppBar({super.key});
-
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        actions: [],
-        elevation: 0,
-        leading: const BackButton(
-        ),
-      ),
-    );
-  }
-
-  @override
-  get preferredSize => throw UnimplementedError();
-}
